+++ CHANGELOG.md
@@ -11,6 +11,11 @@
 * Add `kala.collection.Set::added(E)`
 * Add `kala.collection.Set::addedAll(E...)`
 * Add `kala.collection.Set::addedAll(Iterable<E>)`
+* Add `kala.collection.SetLike::removed(E)`
+* Add `kala.collection.SetLike::removedAll(E...)`
+* Add `kala.collection.SetLike::removedAll(Iterable<E>)`
+* Add `kala.function.Predicates.isNotEquals(T)`
+* Add `kala.function.Predicates.isNotSame(T)`
 
 ### Breaking Changes